implement window functions over already windowed inputs.
