+++ crates/arroyo-df/src/plan/aggregate.rs
@@ -1,0 +1,159 @@
+use crate::extension::aggregate::AggregateExtension;
+use crate::extension::key_calculation::KeyCalculationExtension;
+use crate::plan::WindowDetectingVisitor;
+use crate::{find_window, WindowBehavior};
+use datafusion_common::tree_node::{TreeNode, TreeNodeRewriter};
+use datafusion_common::{DFField, DFSchema, DataFusionError, Result as DFResult};
+use datafusion_expr::{Aggregate, Expr, Extension, LogicalPlan};
+use std::sync::Arc;
+
+#[derive(Debug, Default)]
+pub struct AggregateRewriter {}
+
+impl TreeNodeRewriter for AggregateRewriter {
+    type N = LogicalPlan;
+
+    fn mutate(&mut self, node: Self::N) -> DFResult<Self::N> {
+        let LogicalPlan::Aggregate(Aggregate {
+            input,
+            mut group_expr,
+            aggr_expr,
+            schema,
+            ..
+        }) = node
+        else {
+            return Ok(node);
+        };
+        let mut window_group_expr: Vec<_> = group_expr
+            .iter()
+            .enumerate()
+            .filter_map(|(i, expr)| {
+                find_window(expr)
+                    .map(|option| option.map(|inner| (i, inner)))
+                    .transpose()
+            })
+            .collect::<anyhow::Result<Vec<_>>>()
+            .map_err(|err| DataFusionError::Plan(err.to_string()))?;
+
+        if window_group_expr.len() > 1 {
+            return Err(datafusion_common::DataFusionError::NotImplemented(format!(
+                "do not support {} window expressions in group by",
+                window_group_expr.len()
+            )));
+        }
+
+        let mut key_fields: Vec<DFField> = schema
+            .fields()
+            .iter()
+            .take(group_expr.len())
+            .cloned()
+            .map(|field| {
+                DFField::new(
+                    field.qualifier().cloned(),
+                    &format!("_key_{}", field.name()),
+                    field.data_type().clone(),
+                    field.is_nullable(),
+                )
+            })
+            .collect::<Vec<_>>();
+
+        let mut window_detecting_visitor = WindowDetectingVisitor::default();
+        input.visit(&mut window_detecting_visitor)?;
+
+        let window = window_detecting_visitor.window;
+        let window_behavior = match (window.is_some(), !window_group_expr.is_empty()) {
+            (true, true) => {
+                let input_window = window.unwrap();
+                let (window_index, group_by_window_type) = window_group_expr.pop().unwrap();
+                if group_by_window_type != input_window {
+                    return Err(DataFusionError::NotImplemented(
+                        "window in group by does not match input window".to_string(),
+                    ));
+                }
+                group_expr[window_index] = Expr::Column(window_detecting_visitor.fields.iter().next().unwrap().qualified_column());
+                WindowBehavior::InData
+            }
+            (true, false) => WindowBehavior::InData,
+            (false, true) => {
+                // strip out window from group by, will be handled by operator.
+                let (window_index, window_type) = window_group_expr.pop().unwrap();
+                group_expr.remove(window_index);
+                key_fields.remove(window_index);
+                let window_field = schema.field(window_index).clone();
+                WindowBehavior::FromOperator {
+                    window: window_type,
+                    window_field,
+                    window_index,
+                }
+            }
+            (false, false) => {
+                return Err(DataFusionError::NotImplemented(
+                    format!("must have window in aggregate. Make sure you are calling one of the windowing functions (hop, tumble, session) or using the window field of the input"),
+                ))
+            }
+        };
+
+        let key_count = key_fields.len();
+        key_fields.extend(input.schema().fields().clone());
+
+        let key_schema = Arc::new(DFSchema::new_with_metadata(
+            key_fields,
+            schema.metadata().clone(),
+        )?);
+
+        let mut key_projection_expressions = group_expr.clone();
+        key_projection_expressions.extend(
+            input
+                .schema()
+                .fields()
+                .iter()
+                .map(|field| Expr::Column(field.qualified_column())),
+        );
+
+        let key_projection =
+            LogicalPlan::Projection(datafusion_expr::Projection::try_new_with_schema(
+                key_projection_expressions.clone(),
+                input.clone(),
+                key_schema.clone(),
+            )?);
+
+        let key_plan = LogicalPlan::Extension(Extension {
+            node: Arc::new(KeyCalculationExtension::new(
+                key_projection,
+                (0..key_count).collect(),
+            )),
+        });
+        let mut aggregate_schema_fields = schema.fields().clone();
+        if let WindowBehavior::FromOperator {
+            window: _,
+            window_field: _,
+            window_index,
+        } = &window_behavior
+        {
+            aggregate_schema_fields.remove(*window_index);
+        }
+        let internal_schema = Arc::new(DFSchema::new_with_metadata(
+            aggregate_schema_fields,
+            schema.metadata().clone(),
+        )?);
+
+        let rewritten_aggregate = Aggregate::try_new_with_schema(
+            Arc::new(key_plan),
+            group_expr,
+            aggr_expr,
+            internal_schema,
+        )?;
+
+        let aggregate_extension = AggregateExtension::new(
+            window_behavior,
+            LogicalPlan::Aggregate(rewritten_aggregate),
+            (0..key_count).collect(),
+        );
+        let final_plan = LogicalPlan::Extension(Extension {
+            node: Arc::new(aggregate_extension),
+        });
+        // check that the windowing is correct
+        WindowDetectingVisitor::get_window(&final_plan)?;
+        Ok(final_plan)
+    }
+}
