+++ crates/arroyo-df/src/tables.rs
@@ -700,7 +700,6 @@ impl Insert {
         }
 
         let logical_plan = produce_optimized_plan(statement, schema_provider)?;
-        info!("logical plan: {:#?}", logical_plan);
 
         match &logical_plan {
             LogicalPlan::Dml(DmlStatement {
@@ -710,8 +709,6 @@ impl Insert {
                 ..
             }) => {
                 let sink_name = table_name.to_string();
-                info!("insert logical plan: {:?}", input);
-                info!("insert schema: {:?}", input.schema());
                 Ok(Insert::InsertQuery {
                     sink_name,
                     logical_plan: (**input).clone(),
