+++ crates/arroyo-df/src/test/queries/mastodon_trends.sql
@@ -1,4 +1,3 @@
---fail
 CREATE TABLE mastodon (
     value TEXT
 ) WITH (
