+++ crates/arroyo-df/src/test/queries/no_updating_window_functions.sql
@@ -1,6 +1,3 @@
---fail=SQL window functions are not currently supported
-
--- when window functions are supported, this error should be reported:
---fail=window function must be partitioned by a window as the first argument
+--fail=Window functions require already windowed input
 SELECT *, row_number() OVER (partition by bid.auction order by bid.datetime desc) as row_num
      FROM nexmark where bid is not null
