+++ crates/arroyo-worker/src/arrow/mod.rs
@@ -27,6 +27,7 @@ pub mod session_aggregating_window;
 pub mod sliding_aggregating_window;
 pub(crate) mod sync;
 pub mod tumbling_aggregating_window;
+pub mod window_fn;
 
 pub struct ValueExecutionOperator {
     name: String,
